mass updates

+++ unblock_keenetic.sh
@@ -1,8 +1,56 @@
-#!/bin/sh
+#!/opt/bin/bash
+
+# 8/16 Color vraibles:
+TXT_GRN='\e[0;32m'
+TXT_RED='\e[0;31m'
+TXT_YLW='\e[0;33m'
+TXT_BLUE='\e[0;34m'
+TXT_RST='\e[0m'
+
+WGET='/opt/bin/wget'
+
+github_link='https://raw.githubusercontent.com/elky92'
+
+### Functions for output formatted text
+function echo_OK()
+{
+    echo -e "   ${TXT_GRN}OK${TXT_RST}"
+}
+
+function echo_FAIL()
+{
+    echo -e "   ${TXT_RED}FAIL${TXT_RST}"
+}
+
+function echo_RESULT()
+{
+    local result=$*
+    if [[ "$result" -eq 0 ]]; then
+        echo_OK
+    else
+        echo_FAIL
+        return 1
+    fi
+}
+
+function confirm_reboot()
+{
+    # call with a prompt string or use a default
+    read -r -p "${1:-Are you want to reboot now? [y/N]} " response
+    case "$response" in
+        [yY][eE][sS]|[yY]) 
+            true
+            ;;
+        *)
+            false
+            ;;
+    esac
+}
+
 
 if [ "$1" == "remove" ]
 then
-  opkg remove mc tor tor-geoip bind-dig cron dnsmasq-full ipset iptables dnscrypt-proxy2
+  opkg remove --force-depends --force-removal-of-dependent-packages --autoremove mc tor tor-geoip bind-dig cron dnsmasq-full ipset iptables dnscrypt-proxy2
   rm -rf /opt/etc/ndm/fs.d/100-ipset.sh
   rm -rf /opt/etc/tor/torrc
   rm -rf /opt/etc/unblock.txt
@@ -18,10 +66,9 @@ then
   
   ndmq -p 'no opkg dns-override'
   ndmq -p 'system configuration save'
-  ndmq -p 'system reboot'
+  confirm_reboot && ndmq -p 'system reboot'
   
   sleep 5
-  
   exit 0
 fi
 
@@ -29,27 +76,34 @@ if [ "$1" == "dnscrypt" ]
 then
   if [ ! -f /opt/etc/init.d/S99unblock ]; then
     echo "Ошибка! Основной метод обхода блокировок не реализован в системе. Запустите unblock_keenetic.sh без параметров."
-    exit 0
+    exit 1
   fi
   
   opkg update
   opkg install dnscrypt-proxy2
-  
+  echo_RESULT $?  
+
   rm -rf /opt/etc/dnscrypt-proxy.toml
-  wget --no-check-certificate -O /opt/etc/dnscrypt-proxy.toml https://raw.githubusercontent.com/Kyrie1965/unblock_keenetic/master/dnscrypt-proxy.toml
-  
+  echo -en "$WGET -q --no-check-certificate -O /opt/etc/dnscrypt-proxy.toml $github_link/unblock_keenetic/master/dnscrypt-proxy.toml  ...    "
+  $WGET -q --no-check-certificate -O /opt/etc/dnscrypt-proxy.toml $github_link/unblock_keenetic/master/dnscrypt-proxy.toml  
   /opt/etc/init.d/S09dnscrypt-proxy2 start
-  
+  echo_RESULT $?
+
   rm -rf /opt/bin/unblock_ipset.sh
-  wget --no-check-certificate -O /opt/bin/unblock_ipset.sh https://raw.githubusercontent.com/Kyrie1965/unblock_keenetic/master/unblock_ipset_dnscrypt.sh
+  echo -en "$WGET -q --no-check-certificate -O /opt/bin/unblock_ipset.sh $github_link/unblock_keenetic/master/unblock_ipset_dnscrypt.sh...    "
+  $WGET -q --no-check-certificate -O /opt/bin/unblock_ipset.sh $github_link/unblock_keenetic/master/unblock_ipset_dnscrypt.sh
+  echo_RESULT $?
   chmod +x /opt/bin/unblock_ipset.sh
   
   rm -rf /opt/bin/unblock_dnsmasq.sh
-  wget --no-check-certificate -O /opt/bin/unblock_dnsmasq.sh https://raw.githubusercontent.com/Kyrie1965/unblock_keenetic/master/unblock_dnsmasq_dnscrypt.sh
+  echo -en "$WGET -q --no-check-certificate -O /opt/bin/unblock_dnsmasq.sh $github_link/unblock_keenetic/master/unblock_dnsmasq_dnscrypt.sh...    "
+  $WGET -q --no-check-certificate -O /opt/bin/unblock_dnsmasq.sh $github_link/unblock_keenetic/master/unblock_dnsmasq_dnscrypt.sh
+  echo_RESULT $?
   chmod +x /opt/bin/unblock_dnsmasq.sh
   
   unblock_update.sh
-  
+  echo_RESULT $?
+
   exit 0
 fi
 
@@ -68,6 +122,7 @@ rm -rf /opt/etc/dnscrypt-proxy.toml
   
 opkg update
 opkg install mc tor tor-geoip bind-dig cron dnsmasq-full ipset iptables 
+echo_RESULT $?
 
 set_type="hash:net"
 
@@ -79,47 +134,67 @@ fi
 
 lanip=$(ndmq -p 'show interface Bridge0' -P address)
 
-wget --no-check-certificate -O /opt/etc/ndm/fs.d/100-ipset.sh https://raw.githubusercontent.com/Kyrie1965/unblock_keenetic/master/100-ipset.sh
+echo -en "$WGET -q --no-check-certificate -O /opt/etc/ndm/fs.d/100-ipset.sh $github_link/unblock_keenetic/master/100-ipset.sh...    "
+$WGET -q --no-check-certificate -O /opt/etc/ndm/fs.d/100-ipset.sh $github_link/unblock_keenetic/master/100-ipset.sh
+echo_RESULT $?
 chmod +x /opt/etc/ndm/fs.d/100-ipset.sh
 sed -i "s/hash:net/${set_type}/g" /opt/etc/ndm/fs.d/100-ipset.sh
 
 rm -rf /opt/etc/tor/torrc
-wget --no-check-certificate -O /opt/etc/tor/torrc https://raw.githubusercontent.com/Kyrie1965/unblock_keenetic/master/torrc
+echo -en "$WGET -q --no-check-certificate -O /opt/etc/tor/torrc $github_link/unblock_keenetic/master/torrc...    "
+$WGET -q --no-check-certificate -O /opt/etc/tor/torrc $github_link/unblock_keenetic/master/torrc
+echo_RESULT $?
 sed -i "s/192.168.1.1/${lanip}/g" /opt/etc/tor/torrc
 
-wget --no-check-certificate -O /opt/etc/unblock.txt https://raw.githubusercontent.com/Kyrie1965/unblock_keenetic/master/unblock.txt
+echo -en "$WGET -q --no-check-certificate -O /opt/etc/unblock.txt $github_link/unblock_keenetic/master/unblock.txt...    "
+$WGET -q --no-check-certificate -O /opt/etc/unblock.txt $github_link/unblock_keenetic/master/unblock.txt
+echo_RESULT $?
 
-wget --no-check-certificate -O /opt/bin/unblock_ipset.sh https://raw.githubusercontent.com/Kyrie1965/unblock_keenetic/master/unblock_ipset.sh
+echo -en "$WGET -q --no-check-certificate -O /opt/bin/unblock_ipset.sh $github_link/unblock_keenetic/master/unblock_ipset.sh...    "
+$WGET -q --no-check-certificate -O /opt/bin/unblock_ipset.sh $github_link/unblock_keenetic/master/unblock_ipset.sh
+echo_RESULT $?
 chmod +x /opt/bin/unblock_ipset.sh
 
-wget --no-check-certificate -O /opt/bin/unblock_dnsmasq.sh https://raw.githubusercontent.com/Kyrie1965/unblock_keenetic/master/unblock_dnsmasq.sh
+echo -en "$WGET -q --no-check-certificate -O /opt/bin/unblock_dnsmasq.sh $github_link/unblock_keenetic/master/unblock_dnsmasq.sh...    "
+$WGET -q --no-check-certificate -O /opt/bin/unblock_dnsmasq.sh $github_link/unblock_keenetic/master/unblock_dnsmasq.sh
+echo_RESULT $?
 chmod +x /opt/bin/unblock_dnsmasq.sh
 unblock_dnsmasq.sh
+echo_RESULT $?
 
-wget --no-check-certificate -O /opt/bin/unblock_update.sh https://raw.githubusercontent.com/Kyrie1965/unblock_keenetic/master/unblock_update.sh
+echo -en "$WGET -q --no-check-certificate -O /opt/bin/unblock_update.sh $github_link/unblock_keenetic/master/unblock_update.sh...    "
+$WGET -q --no-check-certificate -O /opt/bin/unblock_update.sh $github_link/unblock_keenetic/master/unblock_update.sh
+echo_RESULT $?
 chmod +x /opt/bin/unblock_update.sh
 
-wget --no-check-certificate -O /opt/etc/init.d/S99unblock https://raw.githubusercontent.com/Kyrie1965/unblock_keenetic/master/S99unblock
+echo -en "$WGET -q --no-check-certificate -O /opt/etc/init.d/S99unblock $github_link/unblock_keenetic/master/S99unblock...    "
+$WGET -q --no-check-certificate -O /opt/etc/init.d/S99unblock $github_link/unblock_keenetic/master/S99unblock
+echo_RESULT $?
 chmod +x /opt/etc/init.d/S99unblock
 sed -i "s/hash:net/${set_type}/g" /opt/etc/init.d/S99unblock
 sed -i "s/192.168.1.1/${lanip}/g" /opt/etc/init.d/S99unblock
 
-wget --no-check-certificate -O /opt/etc/ndm/netfilter.d/100-redirect.sh https://raw.githubusercontent.com/Kyrie1965/unblock_keenetic/master/100-redirect.sh
+echo -en "$WGET -q --no-check-certificate -O /opt/etc/ndm/netfilter.d/100-redirect.sh $github_link/unblock_keenetic/master/100-redirect.sh...    "
+$WGET -q --no-check-certificate -O /opt/etc/ndm/netfilter.d/100-redirect.sh $github_link/unblock_keenetic/master/100-redirect.sh
+echo_RESULT $?
 chmod +x /opt/etc/ndm/netfilter.d/100-redirect.sh
 sed -i "s/hash:net/${set_type}/g" /opt/etc/ndm/netfilter.d/100-redirect.sh
 sed -i "s/192.168.1.1/${lanip}/g" /opt/etc/ndm/netfilter.d/100-redirect.sh
 
 rm -rf /opt/etc/dnsmasq.conf
-wget --no-check-certificate -O /opt/etc/dnsmasq.conf https://raw.githubusercontent.com/Kyrie1965/unblock_keenetic/master/dnsmasq.conf
+echo -en "$WGET -q --no-check-certificate -O /opt/etc/dnsmasq.conf $github_link/unblock_keenetic/master/dnsmasq.conf...    "
+$WGET -q --no-check-certificate -O /opt/etc/dnsmasq.conf $github_link/unblock_keenetic/master/dnsmasq.conf
+echo_RESULT $?
 sed -i "s/192.168.1.1/${lanip}/g" /opt/etc/dnsmasq.conf
 
 rm -rf /opt/etc/crontab
-wget --no-check-certificate -O /opt/etc/crontab https://raw.githubusercontent.com/Kyrie1965/unblock_keenetic/master/crontab
+echo -en "$WGET -q --no-check-certificate -O /opt/etc/crontab $github_link/unblock_keenetic/master/crontab...    "
+$WGET -q --no-check-certificate -O /opt/etc/crontab $github_link/unblock_keenetic/master/crontab
+echo_RESULT $?
 
-ndmq -p 'opkg dns-override'
+#ndmq -p 'opkg dns-override'
 ndmq -p 'system configuration save'
-ndmq -p 'system reboot'
+confirm_reboot && ndmq -p 'system reboot'
 
 sleep 5
-
 exit 0
